# perf: use must-revalidate cache-control header as common and only create header routes for routes with different cache-control