… packages/gatsby/src/utils/adapter/__tests__/__snapshots__/manager.ts.snap