## Description

before on left, with the change in the pr on right for sample reproduction with 10000 dummy pages (this is zoomed in on execution of just this line https://github.com/gatsbyjs/gatsby/blob/94b24825c8811ca968fadade3a0529406bc5390d/packages/gatsby-adapter-netlify/src/route-handler.ts#L262-L263 ):
![image](https://github.com/gatsbyjs/gatsby/assets/419821/ede0418e-5f89-4539-b2da-821243d63435)

Note that in above cpu profile biggest drain is actually way content of `_headers` file is produced and not necessarily amount of them (tho they both combined cause the problem). String concatenations for each route ( https://github.com/gatsbyjs/gatsby/blob/94b24825c8811ca968fadade3a0529406bc5390d/packages/gatsby-adapter-netlify/src/route-handler.ts#L242 ) and then headers concatenation in https://github.com/gatsbyjs/gatsby/blob/94b24825c8811ca968fadade3a0529406bc5390d/packages/gatsby-adapter-netlify/src/route-handler.ts#L136 is causing allocating a lot of new strings and subsequently create a lot of work for garbage collector (and garbage collection is where majority of time was actually spent). This change just lowers amount of times this need to happen, but possibly that part could be improved (i.e. instead of doing `acc += 'string-to-append'` we could push entries to array and finally to `.join('')` on that array to produce final string, but this change alone seems to just mostly avoid hitting issues in the first place

In practice (first reported time is just the build, and second is build + time spent in adapter handling): 
before:
```
info Done building in 25.2874965 sec
✨  Done in 140.97s.
```
and after
```
info Done building in 25.716898292 sec
✨  Done in 33.33s.
```

this also result in much smaller `public/_headers` file as we no longer create entries for each html and page-data files - see https://www.diffchecker.com/VWVHsM5R/

### Documentation

<!--
  Where is this feature or API documented?

  - If docs exist:
    - Update any references, if relevant. This includes Guides and Gatsby Internals docs.
  - If no docs exist:
    - Create a stub for documentation including bullet points for how to use the feature, code snippets (including from happy path tests), etc.
-->

### Tests

Updated unit tests + https://github.com/gatsbyjs/gatsby/blob/master/e2e-tests/adapters/cypress/e2e/headers.cy.ts still passing

## Related Issues

<!--
  Link to the issue that is fixed by this PR (if there is one)
  e.g. Fixes #1234

  Link to an issue that is partially addressed by this PR (if there are any)
  e.g. Addresses #1234

  Link to related issues (if there are any)
  e.g. Related to #1234
-->